Open downloads to a new tab
Open downloads to a new tab instead of browser action.

Browser action now injects JavaScript to the original page to pass
the links for the downloads tab and it opens the downloads tab.

Change content script's global variables from const to var, because
the script can be injected many times. Maybe this could be implemented
in a different way.

The new tab expects the original page to be the previous tab. The new
tab needs the original page's id. This should be fixed somehow.