Change vars back to const in content script
Previously, const functions were changed to vars, because pressing the
browser action again resulted in an error on injecting the content
script.

If you press the browser action second time, without refreshing the
page, content script is injected second time, and it succeeds, because
of the var functions, and it can mess the content script. For example,
when saving the URLs into a file, the URLs in the file are added a
second time etc.
It doesn't matter when injecting the content script the second time, it
fails.