Save URLs by opening a new tab/window
Save selected URLs to a file by opening a new tab or window. Whether a
tab or a window is opened, is up to the user's Firefox preferences. Now,
the downloads page will be left open, but if a new window is opened, it
doesn't close automatically.

TODO
Created object URL should be revoked after use.