Use an iframe when saving URLs to a file
Use an iframe instead of creating a new window or a tab, when saving the
selected URLs to a file. The iframe contains a link, which is pointed to
the file.

This fixes both, the lingering window/tab problem and link not working
sometimes and that the link is saved on the downloads page, blanking the
page.