Try to fix bug #7
Try to fix bug #7 (https://github.com/fluks/download-plenty/issues/7)
by starting the interval and saving the start time and only updating the
status of downloads started after that. When there are no such downloads
in progress, the interval is shut down.

Note that this is probably not foolproof! If something is being
downloaded, the interval won't be cleared and this can affect the bug to
appear again.

TODO
The background donwloading related code should be refactored, at least
after this commit it's a hack.