Ignore executeScript errors
When executeScript is called the second time on the same page, it
returns error and fails when awaited and inside try/catch block. It
seems to work in this commit, but if it tries to execute script in
forbidden pages, such as AMO, it fails silently and download page is
opened, but it there is no content script running.