Try to fix progress not shown bug
Try to fix the bug that sometimes surfaces, when the progress of
downloads, or a download, is not shown. The reason probably was that
Date.now isn't precise (see
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/now#Reduced_time_precision).
If Date.now returns a time before DownloadItem's startTime, progress
monitor wouldn't be started.

Information about downloads is now sent in bulk, not just about one at
the time. This should be more efficient (?).

Clear interval when downloads tab is closed. This is mainly done if the
tab is closed, while downloads are still active, otherwise the inteval
wouldn't be cleared. It also enables to clear the interval by closing the
tab, if the interval isn't cleared for some reason.

Should fix bug #12 (https://github.com/fluks/download-plenty/issues/12).