Fix downloads with same file name bug
Fix a bug when downloading files with the same file name. You can
reproduce the bug by making a Google Images search and downloading the
images. The filename will be "images" for all the images and only one
will succeed. Downloads API has a conflictAction option, but it checks
only existing files, not the ongoing downloads. If there is a file
named "images" on disk before the download is started, a counter is
added to the filenames and all the downloads will succeed.

The fix is to add a counter to filenames if there are files with the
same filenames.